decima instead of uint

+++ LibraryApi/DTOs/Book/BookUpdateDto.cs
@@ -4,7 +4,7 @@
     {
         public int Id { get; set; }
         public string Title { get; set; }
-        public uint Price { get; set; }
+        public decimal Price { get; set; }
         public int AuthorId { get; set; }
         public int GenreId { get; set; }
     }
